+++ src/commands/build.ts
@@ -5,44 +5,7 @@ import { glob } from 'glob'
 import { configDotenv } from 'dotenv'
 import type { TheoreticalEnv } from '../types/config.d.ts'
 import assert from 'node:assert'
-import fs from 'fs/promises'
-import path from 'node:path'
-import { createRequire } from 'node:module'
-const validExtensions = ['.ts','.js', '.json', '.png', '.jpg', '.jpeg', '.webp']
-
-const require = createRequire(import.meta.url)
-//https://github.com/evanw/esbuild/issues/1051
-export const imageLoader = {
-    name: 'attachment-loader',
-    setup: b => {
-        const filter = new RegExp(`\.${validExtensions.slice(3).join('|')}$`)
-        b.onResolve({ filter }, args => {
-            //if the module is being imported, resolve the path and transform to js
-            if(args.importer) {
-               const newPath = path
-                    .format({ ...path.parse(args.path), base: '', ext: '.js' })
-                    .split(path.sep)
-                    .join(path.posix.sep)
-               return { path: newPath, namespace: 'attachment-loader', external: true }
-            }
-            // if the file is actually the attachment, resolve the full dir 
-            return { path: require.resolve(args.path, { paths: [args.resolveDir] }), namespace: 'attachment-loader' }
-        })       
-
-        b.onLoad({ filter: /.*/, namespace: 'attachment-loader' },
-            async (args) => {
-                const base64 = await fs.readFile(args.path).then(s => s.toString('base64'))
-                return {
-                    contents: `
-                    var __toBuffer = (base64) => Buffer.from(base64, "base64");
-                    module.exports = {
-                        name: '${basename(args.path)}',
-                        attachment: __toBuffer("${base64}") 
-                    }`,
-                }
-            })
-    }
-} satisfies Plugin
+import { imageLoader, validExtensions } from '../plugins/imageLoader'
 
 export async function build() {
     const config = await getConfig()
